Updating Auth0 docs
Made changes on the front-end
