+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
+import { BrowserRouter, Route, Switch } from "react-router-dom";
 import { useAuth0 } from "@auth0/auth0-react";
-import history from "./utils/history";
 
 function App() {
   return (
@@ -20,8 +20,6 @@ export default App;
 
 // https://github.com/auth0-samples/auth0-react-samples/tree/master/Sample-01
 
-// import React from "react";
-// import { Router, Route, Switch } from "react-router-dom";
 // import { Container } from "reactstrap";
 
 // import Loading from "./components/Loading";
@@ -52,7 +50,7 @@ export default App;
 //   }
 
 //   return (
-//     <Router history={history}>
+//     <BrowserRouter>
 //       <div id="app" className="d-flex flex-column h-100">
 //         <NavBar />
 //         <Container className="flex-grow-1 mt-5">
